Fix undefined template function error by putting .cpp content to .hpp
- extract `RETRANS_EN` as a CMake parameter, 0 by default
