+++ hw/services/network/hls/ib_transport_protocol/ib_transport_protocol.hpp
@@ -46,8 +46,6 @@ const uint32_t IMMDT_SIZE = 32;
 
 const ap_uint<16> RDMA_DEFAULT_PORT = 0x12B7; //4791 --> 0x12B7
 
-#define RETRANS_EN 1
-
 // QP/EE states, page 473
 typedef enum {RESET, INIT, READY_RECV, READY_SEND, SQ_ERROR, ERROR} qpState;
