+++ hw/services/network/hls/rocev2/CMakeLists.txt
@@ -22,6 +22,7 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/../../cma
 
 
 set(FPGA_PART xcvu11p-flga2577-1-e CACHE STRING "FPGA part")
+set(RETRANS_EN 0 CACHE STRING "Re-transmission logic enable")
 set(DATA_WIDTH 64 CACHE STRING "Width of data path in bytes")
 set(PMTU_BYTES 4096 CACHE STRING "PMTU size.")
 set(CLOCK_PERIOD 6.4 CACHE STRING "Target clock period in nanoseconds")
