+++ hw/services/network/hls/rocev2/make.tcl.in
@@ -13,8 +13,8 @@ add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ipv4/ipv4.cpp -cflags "-I${CMAKE_CURREN
 add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ipv6/ipv6.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
 add_files ${CMAKE_CURRENT_SOURCE_DIR}/../udp/udp.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
 add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ib_transport_protocol/ib_utils.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
-add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ib_transport_protocol/transport_timer.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
-add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ib_transport_protocol/retransmitter/retransmitter.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
+# add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ib_transport_protocol/transport_timer.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
+# add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ib_transport_protocol/retransmitter/retransmitter.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
 add_files ${CMAKE_CURRENT_SOURCE_DIR}/../ib_transport_protocol/ib_transport_protocol.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
 add_files ${CMAKE_CURRENT_SOURCE_DIR}/rocev2.cpp -cflags "-I${CMAKE_CURRENT_BINARY_DIR}"
