+++ hw/services/network/hls/rocev2/rocev2_config.hpp.in
@@ -1,6 +1,8 @@
 #pragma once
 #include <stdint.h>
 
+#define RETRANS_EN ${RETRANS_EN}
+
 const unsigned DATA_WIDTH = ${DATA_WIDTH} * 8;
 
 const uint16_t MAX_QPS = ${ROCE_STACK_MAX_QPS};
